requirements.txt
fastapi
uvicorn
pydantic
requests
streamlit
PyPDF2
python-docx
pandas
matplotlib
openpyxl
fpdf2
textblob
scikit-learn
numpy
chromadb
sentence-transformers
python-multipart
